requirements.txt
firecrawl-py>=0.0.10
playwright>=1.40.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
groq>=0.5.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.24.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
streamlit>=1.28.0
plotly>=5.17.0
wordcloud>=1.9.2
matplotlib>=3.7.0